`Rosstackage` class does probably not necessarily map exactly all the elements defined in `package.xml`, but from the api perspective, it'll be beneficial to be able to access more elements through this class so that client codes won't have to parse xml when needed (here this is related to #65).

Possible downside is that whenever `rospack` crawls it could take longer since more parsing happens (though I haven't tested), and for particularly in usecases where license info is not needed this added info is simply a waste of time. I'm not sure how critical rospack's performance is, so I appreciate opinions.