@dirk-thomas 2016-12-31
Thank you for working on this. Since the performance of `rospack` is very critical it is important to quantify the impact of this patch on that aspect. Can you please run some measurements to compare the runtime with and without this patch on a larger number of packages (>100). Please make sure to compare the results without being impacted by the rospack and other factors (e.g. a different the build type).