@130s 2017-01-21
Ok, I made simple scripts (that I also pushed in https://github.com/ros/rospack/commit/6f3fa0b for sharing purpose) and did small benchmark tests.

Following is the result of running `rospack depends gmapping` 50 times with and without the change in `Rosstackage` if this PR. In a nutshell, 0.5 second slower with the change. 
I'm not sure if this latency is acceptable or not.

With the change:
```
$ git checkout jade-devel
Switched to branch 'jade-devel'
Your branch is up-to-date with 'origin/jade-devel'.
$ git checkout -b j/only_benchmark
Switched to a new branch 'j/only_benchmark'
$ git cherry-pick 6f3fa0b9b5d23b1a6f927848e0cd8f971ff9f2b2
[j/only_benchmark 2a2690f] Add simple benchamrk scripts.
 2 files changed, 67 insertions(+)
 create mode 100755 test/benchmark/iterate_command.sh
 create mode 100755 test/benchmark/measure_performace.sh
$ cd benchmark/
$ git log
commit 2a2690fa9963be5ef1347aa3b7d48edf1b3dd598
Author: Isaac I.Y. Saito <130s@2000.jukuin.keio.ac.jp>
Date:   Sat Jan 21 04:59:49 2017 -0800

    Add simple benchamrk scripts.

commit 205f5bc7da190e7c021956fff6092e932b191a6e
Author: Dirk Thomas <dthomas@osrfoundation.org>
Date:   Fri Sep 2 11:26:39 2016 -0700

    2.3.1
:
$ source ./measure_performace.sh
$ measure_performance_rospack "rospack depends gmapping" 50
Command to be measured: rospack depends gmapping
1th iteration.
cpp_common
rostime
:
tf
2th iteration.
cpp_common
rostime
:
tf
50th iteration.
cpp_common
rostime
:
tf
$ more rospack_50.log 
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 99.99    2.648000       25709       103        52 wait4
  0.01    0.000149           3        51           clone
  0.00    0.000000           0        10           read
  0.00    0.000000           0        50           write
  0.00    0.000000           0        39        30 open
  0.00    0.000000           0        11           close
  0.00    0.000000           0        30        16 stat
  0.00    0.000000           0         8           fstat
  0.00    0.000000           0         3           lseek
  0.00    0.000000           0        15           mmap
  0.00    0.000000           0         8           mprotect
  0.00    0.000000           0         2           munmap
  0.00    0.000000           0        28           brk
  0.00    0.000000           0       111           rt_sigaction
  0.00    0.000000           0       415           rt_sigprocmask
  0.00    0.000000           0        51           rt_sigreturn
  0.00    0.000000           0         1         1 ioctl
  0.00    0.000000           0         9         5 access
  0.00    0.000000           0         1           pipe
  0.00    0.000000           0         1           dup2
  0.00    0.000000           0         1           getpid
  0.00    0.000000           0         1           execve
  0.00    0.000000           0         1           uname
  0.00    0.000000           0         3         1 fcntl
  0.00    0.000000           0         2           getrlimit
  0.00    0.000000           0         5           getuid
  0.00    0.000000           0         5           getgid
  0.00    0.000000           0         5           geteuid
  0.00    0.000000           0         5           getegid
  0.00    0.000000           0         1           getppid
  0.00    0.000000           0         1           getpgrp
  0.00    0.000000           0         1           arch_prctl
------ ----------- ----------- --------- --------- ----------------
100.00    2.648149                   978       105 total
```

Without:
```
$ more rospack_50.log
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 99.99    2.188000       21451       102        51 wait4
  0.01    0.000117           0       415           rt_sigprocmask
  0.00    0.000000           0        10           read
  0.00    0.000000           0        50           write
  0.00    0.000000           0        39        30 open
  0.00    0.000000           0        11           close
  0.00    0.000000           0        30        16 stat
  0.00    0.000000           0         8           fstat
  0.00    0.000000           0         3           lseek
  0.00    0.000000           0        15           mmap
  0.00    0.000000           0         8           mprotect
  0.00    0.000000           0         2           munmap
  0.00    0.000000           0        28           brk
  0.00    0.000000           0       111           rt_sigaction
  0.00    0.000000           0        51           rt_sigreturn
  0.00    0.000000           0         1         1 ioctl
  0.00    0.000000           0         9         5 access
  0.00    0.000000           0         1           pipe
  0.00    0.000000           0         1           dup2
  0.00    0.000000           0         1           getpid
  0.00    0.000000           0        51           clone
  0.00    0.000000           0         1           execve
  0.00    0.000000           0         1           uname
  0.00    0.000000           0         3         1 fcntl
  0.00    0.000000           0         2           getrlimit
  0.00    0.000000           0         5           getuid
  0.00    0.000000           0         5           getgid
  0.00    0.000000           0         5           geteuid
  0.00    0.000000           0         5           getegid
  0.00    0.000000           0         1           getppid
  0.00    0.000000           0         1           getpgrp
  0.00    0.000000           0         1           arch_prctl
------ ----------- ----------- --------- --------- ----------------
100.00    2.188117                   977       104 total
```

```
$ more /proc/cpuinfo 
processor       : 0
vendor_id       : GenuineIntel
cpu family      : 6
model           : 60
model name      : Intel(R) Core(TM) i7-4710MQ CPU @ 2.50GHz
stepping        : 3
microcode       : 0x17
cpu MHz         : 3230.957
cache size      : 6144 KB
physical id     : 0
siblings        : 8
core id         : 0
cpu cores       : 4
:
```